Support Flake8 version 5, drop support for earlier versions.
As of version 5.0.0, due to some refactoring, Flake8 locates and parses
the config files differently. The old hook therefore no longer works
and was replaced with a new one tailored to Flake8 5.0+.

Consequently, we no longer support Flake8 4.x and earlier. But the
dependency declarations should take care of it, so that an older version,
which supported Flake8 4.x, is pip-installed in that case.

One test case, `config_mixed`, had to be modified. It seems that Flake8
will no longer merge configurations from multiple config files, such as
when `pyproject.toml` and `.flake8` and `tox.ini` are all found in the
project folder. We now test that the configuration is read from
`pyproject.toml` in such a scenario, provided it defines a `tool.flake8`
section.