Removed old hook. Rely on plug-in hook from now on.
We scrap the hook introduced only recently, in version 1.0.0, in favor
of the new hook via `flake8.options.config.parse_config()`. It runs
a little later in the process, after another config file may have
already been parsed. We discard that configuration and read from
`pyproject.toml` if we find it in the current folder.

This way, we can rely on the plug-in to load the configuration,
regardless if we're called via `flake8p` or `flake8` itself. The old
hook, via `flake8.options.config._find_config_file()`, would not have
worked like that as Flake8 calls that function before plug-ins are
loaded.