Made test fixtures compatible with Flake8 version 6.0.
Flake8 introduced some breaking changes with the 6.0.0 release that
made parts of the test suite fail. Namely, comments on the same line
as the error code in a config file now raise an error, see:
https://github.com/pycqa/flake8/issues/1689#issuecomment-1326319235

This only affects Flake8's "native" config files in the INI format.
TOML-based config files are not affected.