Don't error out when parsing unknown command-line arguments.
Given who we are, we don't want to parse command-line arguments on
Flake8's behalf. It will do that later in the call chain and report
errors accordingly. We only care if `--toml-config` is present, the
one argument that is "known" to us, but want to defer error reporting
for anything else.